fix: Properly reseve connection for the duration of the request
Problem: connection pool does not properly reserve connection when it is
in use. So, it is possible that two requests within the same session
will start executing in parallel. Mysql driver is not thread-safe and
this will cause malformed network packed.

Fixed by attaching extra mutex to each connection, that would be locked
for duration of the request.